refactoring for additional dynamic loading modules adding features to add/remove collection items
