+++ pylib/apstra/aosom/session.py
@@ -8,21 +8,20 @@ import os
 import datetime
 import socket
 import time
-import importlib
+
 
 from copy import copy
 
 import requests
 import semantic_version
 
+from apstra.aosom.dynmodldr import DynamicModuleOwner
 from apstra.aosom.exc import LoginAuthError, LoginNoServerError, LoginServerUnreachableError
-from apstra.aosom.exc import SessionError
-from apstra.aosom.amods import AosModuleCatalog
 
 __all__ = ['Session']
 
 
-class Session(object):
+class Session(DynamicModuleOwner):
     """
     The Session class is used to create a client connection with the AOS-server.  The general
     process to create a connection is as follows::
@@ -55,7 +54,10 @@ class Session(object):
         * :data:`AOS_PASSWD` - the login user-password, defaults to :data:`~DEFAULTS[\"PASSWD\"]`.
         * :data:`AOS_SESSION_TOKEN` - a pre-existing API session-token to avoid user login/authentication.
     """
-    ModuleCatalog = AosModuleCatalog.keys()
+    DYNMODULEDIR = '.session_modules'
+
+
+    #    ModuleCatalog = AosModuleCatalog.keys()
 
     ENV = {
         'SERVER': 'AOS_SERVER',
@@ -200,18 +202,3 @@ class Session(object):
             # elapsed = datetime.datetime.now() - start
             return False
 
-    # ### ---------------------------------------------------------------------
-    # ###
-    # ###                         DYNAMIC MODULE LOADER
-    # ###
-    # ### ---------------------------------------------------------------------
-
-    def __getattr__(self, amod_name):
-        if amod_name not in AosModuleCatalog:
-            raise SessionError(message='request for unknown module: %s' % amod_name)
-
-        amod_file = AosModuleCatalog[amod_name]
-        got = importlib.import_module(".amods.%s" % amod_file, package=__package__)
-        cls = getattr(got, got.__all__[0])
-        setattr(self, amod_name, cls(api=self.api))
-        return getattr(self, amod_name)
